pyproject.toml
[project]
name = "video-to-website-generator"
version = "0.1.0"
description = "Transform video walkthroughs into functional websites using Gemini and ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "beautifulsoup4>=4.12.0",
    "dirtyjson>=1.0.8",
    "yt-dlp>=2023.12.30",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "aiohttp>=3.10.10"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pyink>=24.3.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
]

[tool.pyink]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "google"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
